refactor: rename http package to httptransport
